mavcomm.jar update, Statusline enhancements, current model display, refactoring scrolling

+++ MAVGCL/src/com/comino/flight/widgets/charts/control/IChartControl.java
@@ -24,7 +24,7 @@ public interface IChartControl {
 
 	public BooleanProperty getCollectingProperty();
 	public IntegerProperty getTimeFrameProperty();
-	public DoubleProperty  getScrollProperty();
+	public IntegerProperty  getScrollProperty();
 
 	public boolean isVisible();
